+++ generator/page_generators/documentation/panther/intrinsics/type_conversion_intrinsics.js
@@ -28,29 +28,29 @@ let page = new Page(__filename, {
 
 
 page.h2Searchable("@bitCast", "bitCast");
-page.text(page.inline_code_block(Language.PANTHER, "func @bitCast = <{FROM: Type, TO: Type}> (from: FROM) -> TO;"));
-page.text(`Convert a value of any type to any other of the same size. Requires that ${page.inline_code_block(Language.PANTHER, "@sizeOf<{FROM}>() == @sizeOf<{TO}>()")}. ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be different types.`);
+page.text(page.inline_code_block(Language.PANTHER, "func @bitCast = <{FROM: Type, TO: Type}> (from: FROM) #unsafe -> TO;"));
+page.text(`Convert a value of any type to any other of the same size. Requires that ${page.inline_code_block(Language.PANTHER, "@numBytes<{FROM, true}>() == @numBytes<{TO, true}>()")}. ${html.inline_code("FROM")} and ${html.inline_code("TO")} cannot be type ${page.inline_code_block(Language.PANTHER, "Void")}.`);
 
 page.h2Searchable("@trunc", "trunc");
 page.text(page.inline_code_block(Language.PANTHER, "func @trunc = <{FROM: Type, TO: Type}> (from: FROM) -> TO;"));
-page.text(`Truncate any ${terms.get("integral")} type to any other smaller ${terms.get("integral")} type. Both ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be ${terms.get("integral")} or a ${terms.get("vector")} of ${terms.get("integral")}. Requires ${page.inline_code_block(Language.PANTHER, "@sizeOf<{FROM}>() > @sizeOf<{TO}>()")}. If is a ${terms.get("vector")}, the lenghts of the vectors must be the same size.`);
+page.text(`Truncate any ${terms.get("integral")} type to any other smaller ${terms.get("integral")} type. Both ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be ${terms.get("integral")} or a ${terms.get("vector")} of ${terms.get("integral")}. Requires ${page.inline_code_block(Language.PANTHER, "@numBits<{FROM, false}>() > @numBits<{TO, false}>()")}. If is a ${terms.get("vector")}, the lengths of the vectors must be the same size.`);
 
 page.h2Searchable("@ftrunc", "ftrunc");
 page.text(page.inline_code_block(Language.PANTHER, "func @ftrunc = <{FROM: Type, TO: Type}> (from: FROM) -> TO;"));
-page.text(`Truncate any ${terms.get("floating-point")} type to any other smaller ${terms.get("floating-point")} type. Both ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be ${terms.get("floating-point")} or a ${terms.get("vector")} of ${terms.get("floating-point")}. Requires the ${terms.get("floating-point")} width of ${html.inline_code("FROM")} to be greater than that of ${html.inline_code("TO")}. If is a ${terms.get("vector")}, the lenghts of the vectors must be the same size.`);
+page.text(`Truncate any ${terms.get("floating-point")} type to any other smaller ${terms.get("floating-point")} type. Both ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be ${terms.get("floating-point")} or a ${terms.get("vector")} of ${terms.get("floating-point")}. Requires ${page.inline_code_block(Language.PANTHER, "@numBits<{FROM, false}>() > @numBits<{TO, false}>()")}. If is a ${terms.get("vector")}, the lengths of the vectors must be the same size.`);
 
 page.h2Searchable("@sext", "sext");
 page.text(page.inline_code_block(Language.PANTHER, "func @sext = <{FROM: Type, TO: Type}> (from: FROM) -> TO;"));
-page.text(`Sign extend any ${terms.get("integral")} type to any other larger ${terms.get("integral")} type. Both ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be ${terms.get("integral")} or a ${terms.get("vector")} of ${terms.get("integral")}. Requires ${page.inline_code_block(Language.PANTHER, "@sizeOf<{FROM}>() < @sizeOf<{TO}>()")}. If is a ${terms.get("vector")}, the lenghts of the vectors must be the same size.`);
+page.text(`Sign extend any ${terms.get("integral")} type to any other larger ${terms.get("integral")} type. Both ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be ${terms.get("integral")} or a ${terms.get("vector")} of ${terms.get("integral")}. Requires ${page.inline_code_block(Language.PANTHER, "@numBits<{FROM, false}>() < @numBits<{TO, false}>()")}. If is a ${terms.get("vector")}, the lengths of the vectors must be the same size.`);
 
 page.h2Searchable("@zext", "zext");
 page.text(page.inline_code_block(Language.PANTHER, "func @zext = <{FROM: Type, TO: Type}> (from: FROM) -> TO;"));
-page.text(`Zero extend any ${terms.get("integral")} type to any other larger ${terms.get("integral")} type. Both ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be ${terms.get("integral")} or a ${terms.get("vector")} of ${terms.get("integral")}. Requires ${page.inline_code_block(Language.PANTHER, "@sizeOf<{FROM}>() < @sizeOf<{TO}>()")}. If is a ${terms.get("vector")}, the lenghts of the vectors must be the same size.`);
+page.text(`Zero extend any ${terms.get("integral")} type to any other larger ${terms.get("integral")} type. Both ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be ${terms.get("integral")} or a ${terms.get("vector")} of ${terms.get("integral")}. Requires ${page.inline_code_block(Language.PANTHER, "@numBits<{FROM, false}>() < @numBits<{TO, false}>()")}. If is a ${terms.get("vector")}, the lengths of the vectors must be the same size.`);
 
 
 page.h2Searchable("@fext", "fext");
 page.text(page.inline_code_block(Language.PANTHER, "func @fext = <{FROM: Type, TO: Type}> (from: FROM) -> TO;"));
-page.text(`Extend any ${terms.get("floating-point")} type to any other larger ${terms.get("floating-point")} type. Both ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be ${terms.get("floating-point")} or a ${terms.get("vector")} of ${terms.get("floating-point")}. Requires the ${terms.get("floating-point")} width of ${html.inline_code("FROM")} to be smaller than that of ${html.inline_code("TO")}. If is a ${terms.get("vector")}, the lenghts of the vectors must be the same size.`);
+page.text(`Extend any ${terms.get("floating-point")} type to any other larger ${terms.get("floating-point")} type. Both ${html.inline_code("FROM")} and ${html.inline_code("TO")} must be ${terms.get("floating-point")} or a ${terms.get("vector")} of ${terms.get("floating-point")}. Requires ${page.inline_code_block(Language.PANTHER, "@numBits<{FROM, false}>() < @numBits<{TO, false}>()")}. If is a ${terms.get("vector")}, the lengths of the vectors must be the same size.`);
 
 page.h2Searchable("@iToF", "iToF");
 page.text(page.inline_code_block(Language.PANTHER, "func @iToF = <{FROM: Type, TO: Type}> (from: FROM) -> TO;"));